Users can delete their account while their orders remain in the database for companies accounting purpose

+++ app/controllers/api/v1/users_controller.rb
@@ -29,7 +29,12 @@ class Api::V1::UsersController < ApplicationController
 
   def destroy
     user = find_user
-    user.destroy
+    user.orders.map do |orderObj|
+      orderObj.update(user_id: 1)
+      if orderObj.valid?
+        user.destroy
+      end
+    end
   end
 
   def profile
